git-svn-id: http://svn.bramp.net:8080/svn/PhD/Work/userspace/threadnetperf@2080 e02701c3-087c-644a-a2a3-551e05f81084

+++ serialise.c
@@ -7,8 +7,8 @@
 #ifdef WIN32
 	typedef signed char             int8_t;
 	typedef unsigned char           uint8_t;
-	typedef signed int              int16_t;
-	typedef unsigned int            uint16_t;
+	typedef signed short            int16_t;
+	typedef unsigned short          uint16_t;
 	typedef signed long int         int32_t;
 	typedef unsigned long int       uint32_t;
 	typedef signed long long int    int64_t;
